Add persistence to Behaviour hooks
Needed to resolve an issue with OrientationBehaviour
where calling getCameraInfo multiple times causes
the transforms to be applied to the same internal object
and cancelling each other out.

Now pass a Bundle to each hook method which should be
unique for each hooked process. Bundle data is not persisted
between process restarts.